added applicationdbcontext and configured connectionstring , added entity in dbset in applicationdbcontext

+++ src/MeetingManagementSystem/MeetingManagementWeb/Models/MeetingMinutesDetails.cs
@@ -3,7 +3,7 @@
     public class MeetingMinutesDetails
     {
         public Guid ServiceProductId { get; set; }
-        public int Quantity { get; set; }
+        public Decimal Quantity { get; set; }
 
 
     }
